research-agent/autonomous_agent/requirements.txt
streamlit>=1.35.0
openai>=1.14.3
dspy>=0.2.2
aiolimiter>=1.0.0
python-dotenv>=1.0.1
requests>=2.31.0
